Refactor to provide option for analyzing a specific route
