+++ frontend/src/components/SectionDropdown.tsx
